feat: fantable adjusted
Starting from 30 because otherwise the daemon wouldn't run when you start the laptop and the original really bad settings would result in a high fan noise. Also the fan duty is adjusted so it stays normally below 30% which is absolutely fine. When your GPU is under load, then it goes up pretty fast, since (at least in my problem setting) the GPU is needed all the time and the temperature needs to be below 94°C for my Nvidia GTX 1070. Also keep in mind that the temperature reading can be a bit delayed so play it safe at high temperatures.